final merging of cosmetic changes to tools.py
